add final changes for 2025
This change is the last for 2025. It introduces two new themes and
a major fix for the Windows installer to properly work with
Powershell 5.1 and above. The Installation summary now provides
more human usable information and missing background errors
no longer polute the run but show up in the summary.